rnotegen_v2/requirements.txt
openai>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
websockets>=11.0.0
dataclasses-json>=0.6.0